Better code documentation
Added comments to the header of the code, with info such as compiler, IDE and their versions being used, among other parameters. Also added a breif description of the application. It was before using a version control, but still may be useful so I'll leave it in the source code